% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glockner2009.R
\docType{data}
\name{glockner2009}
\alias{glockner2009}
\title{Data Set: Multiattributive Decisions (Gloeckner, 2009)}
\format{A data frame 6 variables:
\describe{
  \item{\code{PARTICIPANT}}{Participant number}
  \item{\code{DEC}}{Trial number}
  \item{\code{TYPE}}{Item type (i.e., specific combination of cue values and validities)}
  \item{\code{ACHOICES}}{1 = Option A chosen, 0 = Option B chosen}
  \item{\code{DECTIMES}}{Decision times}
  \item{\code{CONFJUDGMENTS}}{Confidence judgments (between 0 and 100)}
}}
\usage{
glockner2009
}
\description{
Dataset with multiattributive decisions (Glöckner, 2009).
}
\details{
Typical decision experiment.
}
\examples{
head(glockner2009)

}
\references{
Glöckner, A. (2009). Investigating intuitive and deliberate processes statistically: The multiple-measure maximum likelihood strategy classification method. Judgment and Decision Making, 4, 186-199.
}
\keyword{datasets}
