% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luce_choice_rule.R
\name{luce_choice}
\alias{luce_choice}
\title{Luce's Choice Rule}
\usage{
luce_choice(activation, lambda = 2.9)
}
\arguments{
\item{activation}{vector with node activations}

\item{lambda}{parameter used for Luce's choice rule (\link{luce_choice})}
}
\description{
Transforms node-activation to chocie probabilities according to an exponential scaling rule (a special case of Thurstonian scaling).
}
\examples{
luce_choice(c(A=-.5, B=.7, C=.3))
}
\author{
Daniel Heck
}

