% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs_gamble.R
\name{pcs_gamble}
\alias{pcs_gamble}
\title{PCS-DM for Gambles (to do)}
\usage{
pcs_gamble(g1, g2, decay = 0.1, maxiter = 1000, stability = 10^-6,
  convergence = "floor")
}
\arguments{
\item{g1}{Gamble 1, represented as a vector of probabilities, where the possible gains and losses are encoded by the vector names (e.g., \code{g1 = c("100"=.3, "-50"=.2, "0"=.5)})}

\item{g2}{Gamble 2, see \code{g1}}

\item{decay}{decay parameter}

\item{maxiter}{maximum number of iterations}

\item{stability}{stability criterion that determines convergence(i.e., when to stop iterating)}

\item{convergence}{convergence criterion that evaluates the network's energy of the last 10 iterations. One of \code{"floor"} (identical floor of energy/stability for the current vs. remaining 9 nodes), \code{"sum"} (sum of absolute consecutive energy differences), or \code{"max"} (maximum of the absolute differences between the current and remaining 9 nodes)}
}
\description{
Modeling gambles
}

