% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs_multi.R
\name{pcs_multi}
\alias{pcs_multi}
\title{PCS-DM for Multiattribute Binary Decisions}
\usage{
pcs_multi(c1, c2, v, p = 1.9, decay = 0.1, maxiter = 1000,
  stability = 10^-6, convergence = "floor", lambda = 2.9)
}
\arguments{
\item{c1}{cue values for first option (with values +/-1 for positive/negative and 0 for missing cue values)}

\item{c2}{cue values for second option}

\item{v}{vector of cue valdities defined as the probability that cue leads to correct decision if it discriminates between the the options (v=.5 for nonvalid cues)}

\item{p}{sensitivity parameter determining weighting of cue validities (see \link{validity_weight})}

\item{decay}{decay parameter}

\item{maxiter}{maximum number of iterations}

\item{stability}{stability criterion that determines convergence(i.e., when to stop iterating)}

\item{convergence}{convergence criterion that evaluates the network's energy of the last 10 iterations. One of \code{"floor"} (identical floor of energy/stability for the current vs. remaining 9 nodes), \code{"sum"} (sum of absolute consecutive energy differences), or \code{"max"} (maximum of the absolute differences between the current and remaining 9 nodes)}

\item{lambda}{parameter used for Luce's choice rule (\link{luce_choice})}
}
\description{
PCS-DM for Multiattribute Binary Decisions
}
\examples{
bb <- pcs_multi(c1 = c(-1,1,1), c2 = c(1,-1,-1),
               v = c(.85,.75,.6))
bb[1:4]
}

