% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox_multi.R
\name{toolbox_multi}
\alias{toolbox_multi}
\title{Adaptive-Toolbox Strategies for Multiattributive Decisions}
\usage{
toolbox_multi(c1, c2, v, error = 0)
}
\arguments{
\item{c1}{cue values for first option (with values +/-1 for positive/negative and 0 for missing cue values)}

\item{c2}{cue values for second option}

\item{v}{vector of cue valdities defined as the probability that cue leads to correct decision if it discriminates between the the options (v=.5 for nonvalid cues)}

\item{error}{fixed error probability of applying a strategy (usefull for maximum-likelihood estimation)}
}
\description{
For modeling strategies such as take-the-best (TTB), equal-weight (EQW), and weighted-additive (WADD).
}
\examples{
toolbox_multi(c1 = c(1,-1,1,1), c2=c(-1,1,1,1), v=c(.9,.8,.7,.6))
}
\seealso{
For PCS-DM predictions see \link{pcs_multi}
}

