% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_weight.R
\name{validity_weight}
\alias{validity_weight}
\title{PCS-DM Weighting of Validities}
\usage{
validity_weight(v, p = 1.9)
}
\arguments{
\item{v}{validities}

\item{p}{sensitivity parameter}
}
\description{
Transform validities (= probability of cue leading to correct decision) into decision weights for PCS network.
}
\details{
The weighting function is:

\eqn{w = sign(v-.5)*abs(v-.5)^p}

Results in rather compensatory \eqn{p<1} or noncompensatory \eqn{p>1} weighting of cues.
}
\examples{
# compensatory:
curve(validity_weight(x, p=.5), xlab="Validity")
# noncompensatory:
curve(validity_weight(x, p=2.5), xlab="Validity")
}
\author{
Daniel Heck
}
